used separate input_dir for animations
I put a feature where pressing dodge while
standing still would give a value to
input_direction to make the legs move a bit
while dodging. But this caused a problem
with the actual dodge movement as now
that were a move forward 'input' and would
move in the direction of the camera instead
of just where the model is facing. This
commit fixes that by having a separate input
direction variable i give to the anim script.