duct tape fix first attack transitioning to itself
In a blend tree, a transition request to a node
already playing causes it to reset the animation
without a transition, making it instant. This
is not the desired behaviour and does not look
good. Checking the 'Allow Transition to Self'
on the Transition node or toggling the 'Reset'
checkbox does absolutely nothing.

This dumb fix creates a copy of the first
attack and switches between the two attacks to
make it seem like the same animation is playing.

I don't know if this is an issue with Godot or
my poor implementation, but what a dumb problem.