fix attempt enemy going through player after death
problem: turned off collisions of the enemy so that it
can't be interacted with after death. But the behaviour
tree would make it so it would 'attack' and so move
forwards after death.

I think i got this fixed by just setting the flag that allows
for it to attack to be false