hopefully dizzy finisher improvements
+ correct behaviour coming out of finisher
both from parry or damage and either
performing the finisher or letting it go
+ dizzy from damage: can make multiple
enemies dizzy with a hit, but finishing off
an enemy means surrounding enemies will
not recognise being hit

this commit was a doozy. lots of conditions
to be aware of. head scratching was needed
as fixing one problem would cause another
to pop up.