hopefully fix instantly finishing dizzy damage
because i constantly set can damage to true in process.
this was to fix the problem of can damage being set to
false somewhere else. but now since it is true every
frame, it could finish the enemy in the first few frames.
now i just connect to the signal that specified when
it can do damage and then set it to true when received.