improvement movement and dodging
i had a revelation in bed that changing each
component separately meant they changed
independantly instead of affecting the entire
desired movement vector as a whole. this is
why there was weird behaviour when i tried
to move_toward on the x and z component
separately.

changed the dodge behaviour accordingly.