added zombie smack attack
! bro for some reason renaming the blender
animation to Zombie Smack and then saving
the animation to
zombie_animations_1_smack.res causes an
error for add_animation_library
apparently the animation library is null??

like what?>!!? how tf does changing a couple
names screw it up? sorry for the rant it's just
so weird and I've spent hours on it and didnt
even solve it. im so damn hungry.
we both know its a smack but
godot wants to call it a punch...

probably my fault for staying on 4.1