support multiple weapons
only certain weapons can deal damage
during an animation

store dictionary of StringName for weapon
names and NodePath for the Weapon node.

calls to receive_can_damage in the animation
tracks specify a weapon name, that specific
weapon gets affected.

if no weapon name is provided or no name
matches in the dictionary, just affect
all weapons in the dictionary.

also renamed all references to sword to be
weapon instead.

! getting hit twice from the same attack
! still need to improve can damage timings