simple attack strategy
assumes that animation tree has snake
case input for the Attack transition node.

assumes that each transition input is their
own blend tree with a trim and speed
node with {attack_name} Trim/Speed
as their node names.

implemented a has_copy bool but i dont know
if it actually works as intended. supposed to be
for anims in the animation tree that you didn't
make a copy for.