decoupled character animations from enemy
individual components shouldnt be referencing animations
through the parent CharacterAnimations script right?
Just reference them directly when needed which allows
CharacterAnimations to not require storing all animations.

made PlayerAnimations because this stays pretty static
and is a good place for the states to access animations.

The enemy is made up of components enough that they
don't need a dedicated animations script. This may change of
course lmao. Lets see how this goes...