configurable health damage, instability, knockback
closes #61

+ DamageAttributes resource
+ SecondaryMovement resource

DamageSource now has a DamageAttributes resource
that keeps track of health damage, instability on hit,
block, parry, and got parried interactions, and knockback.

Any logic handling incoming damage sources just
accesses the damage attributes field.

Melee attacks can behave differently, so each MeleeAttack
stores a DamageAttributes field. The MeleeAnimations sends
a signal to the MeleeComponent to update the
DamageAttributes field in the weapons.

? right now it just updates the DamageAttributes of all
weapons. should there be a keyframe specified in an
animation track that specifies which weapon to update?

? That also means that animations tracks should be able to
update the DamageAttributes values. This is going under
the assumption that there will be animations with multiple
opportunities for damage. Maybe multiple times with the
same weapon, multiple weapons at the same time, or a mix.
I'LL FIGURE IT OUT LATER.

Once the MeleeComponent updates the DamageAttributes
on the DamageSource, any logic handling the incoming weapon
just accesses the DamageAttributes resource.

It works slightly differently for projectiles. The melee attack
animations are responsible for the damage attributes on the
damage sources (weapons). But for non melee actions, the
animations are independant from the damage sources
(projectiles).

A melee weapon is pretty much an object that gets swung
around, so the animations dictate it's damage. A projectile
is summoned, it configures itself because the animation
basically says "a non melee action effect happens here!"
and it can be anything at all.

The DamageAttributes on each projectile (which extends
DamageSource) are good as is. They can be configured
directly.

So, to reiterate, MeleeAttacks (each animation) have their
own DamageAttributes. The MeleeAnimations takes these
and sends them to the MeleeComponent, which then
updates the DamageAttributes of the weapons
(DamageSources) to those values.

Each NonMeleeAction animation just acts normally. The
NonMeleeAnimations takes the keyframes and sends
signals to the NonMeleeComponent, which then performs
a NonMeleeActionEffect, which then creates projectile(s).
This projectile is a DamageSource so the DamageAttributes
can be modified directly.

DamageAttributes are accessed through the incoming
DamageSource.

Secondary movement now uses a SecondaryMovement
resource. This was done to enable configuration of
knockback in DamageAttributes.